refactoring zipdownload to two components
